Phase-IV/backend/requirements.txt
# Core dependencies
typing-extensions>=4.8.0

# FastAPI and web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database and ORM
sqlmodel==0.0.22
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Authentication and security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==1.0.1

# Validation and utilities
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0

# Phase III: AI and MCP Server
mcp>=1.0.0  # Official Python MCP SDK
google-generativeai>=0.3.0  # Google Gemini API
openai-agents>=0.1.0  # OpenAI Agents SDK

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0  # For async test client
